+++ spec/Pim/Bundle/BaseConnectorBundle/Processor/ProductToFlatArrayProcessorSpec.php
@@ -3,7 +3,7 @@
 namespace spec\Pim\Bundle\BaseConnectorBundle\Processor;
 
 use PhpSpec\ObjectBehavior;
-use Pim\Bundle\CatalogBundle\Entity\Channel;
+use Pim\Bundle\CatalogBundle\Model\ChannelInterface;
 use Pim\Bundle\CatalogBundle\Manager\ChannelManager;
 use Pim\Bundle\CatalogBundle\Model\ProductInterface;
 use Pim\Bundle\CatalogBundle\Model\ProductMediaInterface;
@@ -17,7 +17,7 @@ class ProductToFlatArrayProcessorSpec extends ObjectBehavior
     }
 
     function it_returns_flat_data_with_media(
-        Channel $channel,
+        ChannelInterface $channel,
         $channelManager,
         ProductInterface $item,
         ProductMediaInterface $media1,
@@ -49,7 +49,7 @@ class ProductToFlatArrayProcessorSpec extends ObjectBehavior
     }
 
     function it_returns_flat_data_without_media(
-        Channel $channel,
+        ChannelInterface $channel,
         ChannelManager $channelManager,
         ProductInterface $item,
         Serializer $serializer